+++ src/php/Lang/Collections/HashSet/TransientHashSetInterface.php
@@ -5,19 +5,16 @@ declare(strict_types=1);
 namespace Phel\Lang\Collections\HashSet;
 
 use Countable;
+use Phel\Lang\ContainsInterface;
 
 /**
  * @template V
  *
  * @extends AsTransientInterface<TransientMapInterface>
+ * @extends ContainsInterface<V>
  */
-interface TransientHashSetInterface extends Countable
+interface TransientHashSetInterface extends Countable, ContainsInterface
 {
-    /**
-     * @param V $value
-     */
-    public function contains($value): bool;
-
     /**
      * @param V $value
      */