+++ tests/php/Unit/Lang/Collections/Map/PersistentArrayMapTest.php
@@ -18,8 +18,8 @@ class PersistentArrayMapTest extends TestCase
         $h = PersistentArrayMap::empty(new ModuloHasher(), new SimpleEqualizer());
 
         self::assertEquals(0, $h->count());
-        self::assertFalse($h->containsKey('test'));
-        self::assertFalse($h->containsKey(null));
+        self::assertFalse($h->contains('test'));
+        self::assertFalse($h->contains(null));
         self::assertNull($h->find('test'));
     }
 
@@ -36,10 +36,10 @@ class PersistentArrayMapTest extends TestCase
 
         self::assertEquals(null, $h->find(null));
         self::assertEquals(0, $h->count());
-        self::assertFalse($h->containsKey(null));
+        self::assertFalse($h->contains(null));
         self::assertEquals('test', $h2->find(null));
         self::assertEquals(1, $h2->count());
-        self::assertTrue($h2->containsKey(null));
+        self::assertTrue($h2->contains(null));
     }
 
     public function testPutKeyValue(): void
@@ -48,7 +48,7 @@ class PersistentArrayMapTest extends TestCase
             ->put(1, 'test');
 
         self::assertEquals(1, $h->count());
-        self::assertTrue($h->containsKey(1));
+        self::assertTrue($h->contains(1));
         self::assertEquals('test', $h->find(1));
     }
 
@@ -59,7 +59,7 @@ class PersistentArrayMapTest extends TestCase
             ->put(1, 'test');
 
         self::assertEquals(1, $h->count());
-        self::assertTrue($h->containsKey(1));
+        self::assertTrue($h->contains(1));
         self::assertEquals('test', $h->find(1));
     }
 
@@ -70,7 +70,7 @@ class PersistentArrayMapTest extends TestCase
             ->put(1, 'foo');
 
         self::assertEquals(1, $h->count());
-        self::assertTrue($h->containsKey(1));
+        self::assertTrue($h->contains(1));
         self::assertEquals('foo', $h->find(1));
     }
 
@@ -81,7 +81,7 @@ class PersistentArrayMapTest extends TestCase
             ->put(null, 'test');
 
         self::assertEquals(1, $h->count());
-        self::assertTrue($h->containsKey(null));
+        self::assertTrue($h->contains(null));
         self::assertEquals('test', $h->find(null));
     }
 
@@ -117,7 +117,7 @@ class PersistentArrayMapTest extends TestCase
             ->remove(null);
 
         self::assertEquals(0, $h->count());
-        self::assertFalse($h->containsKey(null));
+        self::assertFalse($h->contains(null));
         self::assertNull($h->find(null));
     }
 
@@ -127,7 +127,7 @@ class PersistentArrayMapTest extends TestCase
             ->remove(null);
 
         self::assertEquals(0, $h->count());
-        self::assertFalse($h->containsKey(null));
+        self::assertFalse($h->contains(null));
         self::assertNull($h->find(null));
     }
 
@@ -137,7 +137,7 @@ class PersistentArrayMapTest extends TestCase
             ->remove(1);
 
         self::assertEquals(0, $h->count());
-        self::assertFalse($h->containsKey(1));
+        self::assertFalse($h->contains(1));
         self::assertNull($h->find(1));
     }
 
@@ -148,9 +148,9 @@ class PersistentArrayMapTest extends TestCase
             ->remove(1);
 
         self::assertEquals(1, $h->count());
-        self::assertTrue($h->containsKey(2));
+        self::assertTrue($h->contains(2));
         self::assertEquals('test', $h->find(2));
-        self::assertFalse($h->containsKey(1));
+        self::assertFalse($h->contains(1));
         self::assertNull($h->find(1));
     }
 
@@ -161,7 +161,7 @@ class PersistentArrayMapTest extends TestCase
             ->remove(1);
 
         self::assertEquals(0, $h->count());
-        self::assertFalse($h->containsKey(1));
+        self::assertFalse($h->contains(1));
         self::assertNull($h->find(1));
     }